chore: split down relationship.test.ts
