+++ test/transforms.test.ts
@@ -12,6 +12,34 @@ test('order', async () => {
       "count": null,
       "data": Array [
         Object {
+          "age_range": null,
+          "catchphrase": null,
+          "data": null,
+          "status": "ONLINE",
+          "username": "testuser3",
+        },
+        Object {
+          "age_range": null,
+          "catchphrase": null,
+          "data": null,
+          "status": "ONLINE",
+          "username": "testuser2",
+        },
+        Object {
+          "age_range": null,
+          "catchphrase": null,
+          "data": null,
+          "status": "ONLINE",
+          "username": "testuser1",
+        },
+        Object {
+          "age_range": null,
+          "catchphrase": null,
+          "data": null,
+          "status": "OFFLINE",
+          "username": "testuser",
+        },
+        Object {
           "age_range": "[1,2)",
           "catchphrase": "'cat' 'fat'",
           "data": null,
@@ -74,13 +102,6 @@ test('order on multiple columns', async () => {
         Object {
           "channel_id": 3,
           "data": null,
-          "id": 3,
-          "message": "Some message on channel wihtout details",
-          "username": "supabot",
-        },
-        Object {
-          "channel_id": 3,
-          "data": null,
           "id": 4,
           "message": "Some message on channel wihtout details",
           "username": "supabot",
@@ -93,10 +114,73 @@ test('order on multiple columns', async () => {
           "username": "supabot",
         },
         Object {
+          "channel_id": 2,
+          "data": null,
+          "id": 3,
+          "message": "foo",
+          "username": "supabot",
+        },
+        Object {
+          "channel_id": 1,
+          "data": null,
+          "id": 13,
+          "message": "test1",
+          "username": "supabot",
+        },
+        Object {
+          "channel_id": 1,
+          "data": null,
+          "id": 14,
+          "message": "test1",
+          "username": "supabot",
+        },
+        Object {
+          "channel_id": 1,
+          "data": null,
+          "id": 15,
+          "message": "updated",
+          "username": "supabot",
+        },
+        Object {
+          "channel_id": 1,
+          "data": null,
+          "id": 16,
+          "message": "test3",
+          "username": "supabot",
+        },
+        Object {
+          "channel_id": 1,
+          "data": null,
+          "id": 17,
+          "message": "test3",
+          "username": "supabot",
+        },
+        Object {
           "channel_id": 1,
           "data": null,
           "id": 1,
           "message": "Hello World 👋",
+          "username": "supabot",
+        },
+        Object {
+          "channel_id": 1,
+          "data": null,
+          "id": 18,
+          "message": "test3",
+          "username": "supabot",
+        },
+        Object {
+          "channel_id": 1,
+          "data": null,
+          "id": 11,
+          "message": "foo",
+          "username": "supabot",
+        },
+        Object {
+          "channel_id": 1,
+          "data": null,
+          "id": 12,
+          "message": "test1",
           "username": "supabot",
         },
       ],
@@ -150,10 +234,17 @@ test('range', async () => {
         },
         Object {
           "age_range": "[20,30)",
-          "catchphrase": "'fat' 'rat'",
-          "data": null,
+          "catchphrase": "'json' 'test'",
+          "data": Object {
+            "foo": Object {
+              "bar": Object {
+                "nested": "value",
+              },
+              "baz": "string value",
+            },
+          },
           "status": "ONLINE",
-          "username": "dragarcia",
+          "username": "jsonuser",
         },
       ],
       "error": null,
@@ -285,8 +376,12 @@ test('csv', async () => {
     supabot,,\\"[1,2)\\",ONLINE,\\"'cat' 'fat'\\"
     kiwicopple,,\\"[25,35)\\",OFFLINE,\\"'bat' 'cat'\\"
     awailas,,\\"[25,35)\\",ONLINE,\\"'bat' 'rat'\\"
+    jsonuser,\\"{\\"\\"foo\\"\\": {\\"\\"bar\\"\\": {\\"\\"nested\\"\\": \\"\\"value\\"\\"}, \\"\\"baz\\"\\": \\"\\"string value\\"\\"}}\\",\\"[20,30)\\",ONLINE,\\"'json' 'test'\\"
     dragarcia,,\\"[20,30)\\",ONLINE,\\"'fat' 'rat'\\"
-    jsonuser,\\"{\\"\\"foo\\"\\": {\\"\\"bar\\"\\": {\\"\\"nested\\"\\": \\"\\"value\\"\\"}, \\"\\"baz\\"\\": \\"\\"string value\\"\\"}}\\",\\"[20,30)\\",ONLINE,\\"'json' 'test'\\"",
+    testuser,,,OFFLINE,
+    testuser1,,,ONLINE,
+    testuser2,,,ONLINE,
+    testuser3,,,ONLINE,",
       "error": null,
       "status": 200,
       "statusText": "OK",
