--- test/types/default-inference.test-d.ts
@@ -1,3 +1,0 @@
-const REST_URL = 'http://localhost:3000'
-
-// Check for PostgrestClient without types provided to the client
